trading-terminal/requirements.txt
flask
flask-cors
gunicorn
aiohttp
pandas
numpy
pytz
python-dotenv
pyjwt
